Starting Code
Clint, I managed to get a pretty good start. There is a problem through with the text file. I read something on stack overflow how windows can do some weird stuff when making a new line in a txt file and this causes the name of the objects inside of the vending machine to be partially over written by the quantity of the item. I want you to remake the txt file. Maybe even using your mac OS. I had this problem in a previous assignment and the new txt file that you sent me seemed to fix the problem.